test_vmware_alarm_selection_does_not_fail moved to test_bugs.py
